use a different way of importing version, PROJECT_NAME etc
A regular import of the submodules of the electroncash would cause the
__init__ of the package to be executed, which imports electroncash.bitcoin,
which raises an error if the python interpreter is called with the -O option.

The tools using setup.py (pip...) may be using `python -O`.

Besides, it is better that the setup does not import the package that it is supposed
to build and install. Currently I works because it is pure-python code. But in the
future we might use compiled extensions which would make the source directory not
importable without a proper build.