copy the Electron Cash data dir if it exists on the first execution
Splitting coins and loading pre-fork wallets will be the first major use cases.
This automates the process of importing the Electron Cash wallets the first time
Electrum ABC is launched..

Test Plan:

```
mv ~/.electrum-abc ~/.electrum-abc-bck
./electrum-abc
```
Check that we see the transactions from the EC wallets if we do not have our own data folder.

Restore the original data, check that this time no data is copied.