enable update checking again
Set the version number to 4.3.0 (remove b0).
Add a new key for signing version numbers.
Sign the version string "4.3.0" with the new key.
Enable various widgets allowing to run update-checking.
Set the download url displayed in the update notification dialog.

Test Plan:
After this commit is pushed to github, downgrade the version number locally to 4.2.0,
and check that it triggers an update notification. Then discard this change (revert the version back to 4.3.0).