change recommended install method to `pip --user`
`sudo pip` is not recommended as it makes uninstalling the package difficult, and it is difficult to predict whether the installation path takes higher or lower priority on the `sys.path` list, compared to system packages (e.g. installed via `sudo apt-get install...`).
User packages generally take priority on system packages, and can be uninstalled and updated without sudo privileges.