improve the code clarity
This commit gives a better name to variables, removes a monkey patch (self.addChild = self.addTopLevelItem), adds a few comments and refactors the method a little bit.

Test Plan:
To test both main branches, run the program on the mainnet, where we have two chains and two sets of servers,  and on the testnet where there is a single chain. Go to the network overview menu and check everything still works.