add a checkpoint for the BCHA chain
This will remove the risk of spending BCH accidentaly after modifying the server settings.

Test Plan:
Launch electrum-abc, try to connect to a BCH electrum server, check that it fails. You should also see the list of banned servers be populated with the BCH servers that your selected server is connected to.