switch to a default derivation path of m/44'/899'/0'
Plan to use m/44'/1899'/0' for SLPA tokens

New wallets are created with this new derivation path, and the path is suggested as a default when restoring a wallet from seed. When pairing with a hardware wallet, keep recommending the Bitcoin Cash derivation path for now, until we get confirmation that hardware wallets support BCHA.

The dialogs still mention the possible derivation paths for BTC and BCH to help users restore old wallets.

Test Plan:
Create a new wallet, check the derivation path in the wallet file, restore it from seed,
test pairing with a Ledger Nano S and a Trezor T.