Overhaul of the "crash report" dialog
Do not use an intermediary server. Send the user directly to github to
open an issue. The issue is prefilled (title, tags, body) if the user
already has a github account. If he needs to create an account first,
he will have to come back to the dialog and copy paste the report
or click again after creating an account.

Remove some convoluted code, such as the keeping of a self reference
by the widget to not disappear immediately after its creation.
It is simpler to let the exception hook keep a reference to the
widget to keep it alive.

Test plan:
Add an intentional error somewhere in a menu action (`raise IOError("test"`)
enable the crash reporter in the menu Tools -> Preferences -> Fees and Misc,
trigger the error and test that the github issue is open and prefilled in your
web browser.