fix a case insensitivity issue on Windows
When replacing the electron Cash data path with the Electrum ABC data
path on Windows, we need to take into account that Windows paths are
case insensitive.

Test Plan:

On a windows computer, delete the Electrum ABC user directory, restart the program, see that now the wallet paths in the config file are properly updated to point to the Electrum ABC data directory.