support all 3 address formats (legacy, CashAddr, CashAddr BCH)
This mega-commit adds support for CashAddr and CashAddr BCH throughout the UI.
CashAddr BCH cannot be entirely dropped until hardware wallets support the new
prefix (altough we could maybe just use legacy for hardware wallets, but
this requires more testing).

A new icon is added for the status bar button that cycles through the address
formats. The icon is red for BCH CashAddr, and we re-use the blue icon for the
new ecash CashAddr.

In the configuration file, the use_cashaddr boolean is replaced with a
address_format string.

Test plan:
Cycle through address types using:
- the dedicated status bar icon
- the combobox in the settings menu

Send a few transactions.

Open a hardware wallet (ledger Nano S), send a transaction.