lower the default fee rate
The new slider values range from 5000 to 50000 sat/kB, instead
of 10000--100000.

The fee is automatically updated in the config file for users upgrading
from older versions of Electrum ABC if they are using the default setting
for their old version.

For users with a custom slider setting that is higher than the new slider
maximum, the fee will be set to the new maximum.