Build: Verify the Python checksums
Before this change, whoever has a copy of the Python release signing key
is a trusted third party. This commit changes this so that in addition
to checking the signature of the installers, the checksums are also
verified.

This is motivated by the Python 3.8.9 release (2021-04-02) being
replaced with a different version (2021-04-06) that didn't match the
PGP signature.