Make PaymentRequest.has_expired always return a boolean (closes #2298)
Backport note: this is backported only to make future backports easier,
and for the unlikely event that someone want to maintain the Android
version of Electrum ABC in the future.
It does not solve any python problem, only a java issue because it does
not seem to cast integers to booleans.