CashFusion: Editing the fuse_depth setting fix for send tab
We had a bug where editing the fuse depth setting for 1 wallet would
affect the way the send tab displayed the text for "Spend only fused
coins" for *all* wallets!

Since this is a per-wallet setting we should only update the checkbox
text for the wallet we are editing.

This commit corrects the situation.