don't show Bug Report window when loading external plugins
This was handled correctly on plugin installation, but after installation the failing plugin exists (disabled) in the list, and the user can click "Enable", which would not catch the error again. As a result the user would see the "Report bug" window encouraging him to open an issue on github.

After this change, the user just sees a generic error message suggesting that he should contact the plugin developers.

TODO? Needs more thinking.

 - uninstall (delete from user data directory) plugins if they failed installing
 - disable external plugins altogether, because we will keep improving and modifying the codebase and do not want compatibility issues with third party code slowing us down.