add a flake8 plugin: simplify
See https://pypi.org/project/flake8-simplify/

SIM106 ignored because it does not simplify the code in the one example that popped up:
```
if A:
   ...
elif B:
   ...
elif C:
   ...
else:
   raise ...
```

I can see how it would make sense if there was a single condition, but with multiple conditions it makes me repeat the conditions and does not get rid of the indentation.

SIM119 (Use dataclasses for data containers) because if has false positives.