fix intermittent test_wallet failure
`TestCreateRestoreWallet.test_create_new_wallet` sometimes fails with a `keystore.InvalidSeed` error. The failure is rare, so there is no way to be sure, but the most likely reason is that the test causes the creation of a bip39 mnemonic phrase by specifying `seed_type=None` (default behavior), and that seed is later wrongly detected as an "electrum" seed in `keystore.from_seed`.

Fix it by specifying the seed type explicitely in the case `create_new_wallet` is called without `seed_type`.

This bug could only affected the test and command line use. `create_new_wallet` is not used anywhere else in the codebase. The GUI does not use that abstraction level and calls mnemo.make_bip39_words directly in base_wizard.py.