remove more occurences of "Bitcoin Cash"
These are mostly reffering to "Bitcoin Cash addresses", but there was also one occurence left in the warning message that is displayed when opening a watch-only wallet.
The error message about trustedcoin wallets no supporting Bitcoin Cash is from 2018, and the link to the fix was wrong anyway, so just remove that altogether.