refactor AddressConsolidator.get_unsigned_transactions to avoid invalidating the object
Prior to this change, the instance of  AddressConsolidator would be invalid after the first call to get_unsigned_transactions.
Fix this by not consuming self._coins, but rather just increment an index to access the coins.

If the instace can be reused, we can avoid doing the initial coin filter again if only the max_tx_size parameter is changed.

Test: `python -m electroncash.tests.test_consolidate`