fix maximum value for minimum amount selection
A QDoubleSpinBox has a default range of 0.00 to 100.00. The maximum value needs to be increased by a lot to make this widget useful.

Set the default minimum value to the dust threshold (5.46 XEC). Users can still set the value below that limit (the minimum remains 0 for the widget), to be future proof in case the dust threshold ever changes.

Change also the step type to `AdaptiveDecimalStepType`. Adaptive decimal step means that the step size will continuously be adjusted to one power of ten below the current value
This works for PyQt5 >= 5.12, and we have a minimum requirement of PyQt5 > 5.15, so it should be fine.