implement Proof deserialization and add tests
Being able to unserialize proofs will be useful for testing or to allow users to specify an external proof when creating a delegation.

Also rename the test module to something more generic, so that we can later group tests for various avalanche modules.
Add some helper generator to build some objects from a hex string, and implement a __str__ to help with debugging a failing test