suggest a proof master key in the avalanche proof tool
This pre-fills the master key field in the tool, so the user does not need to find a key from an external source and he is less likely to just lazily reuse a key from an address that is used in the wallet.

This key is generated deterministically from the wallet seed phrase, but from a BIP44 path that will never be used by the wallet (change_index = 2).
For now, we always use the key at index (2, 0), but in the future we will likely save generated proofs as wallet data so we can start generating different unused keys for each new proof).