make the first page of the wallet wizard more explicit
A user's first contact with the application is this install wizard. It is not immediately clear to some new users that they are creating a new wallet on this screen.

Add a label to explain what is expected of the user. Add tooltips to the various input fields and buttons.

This same wizard is also used to create an additional wallet or load an existing wallet, so the instructions must remain generic enough (no chance to add a "Welcome new user!" message without some additional work, unfortunately).

Closes https://github.com/Bitcoin-ABC/ElectrumABC/issues/189