[avalanche] warn if a coin is below the proof dust threshold
Show a warning if a user attempts the generate a proof with coins below the proof dust threhold. Users can decide to ignore the warning, if they just want to try out the proof generation tool.

Also show amounts in red in the utxo list, if a value is too low.