[avalanche] exclude unconfirmed coins for creating avalanche proofs
Addresses https://github.com/Bitcoin-ABC/ElectrumABC/issues/195
The stake serialization requires a block height, so coins that are not yet included in a block cannot be used.

Electrum assigns height 0 to unconfirmed coins, and height -1 to unconfirmed coins with unconfirmed parents, which caused a bug when trying to serialize the height as an unsigned int. An assertion is added in the Stake constructor to ensure this does not happen.
This has a side effect of making it impossible to build a proof using coins from the genesis block. Sorry, Hal.

In the coins tab, the "Build avalanche proof" action is no longer enabled if an unconfirmed coin is selected. But we still need to deal with the case of unconfirmed coins in the bproof building dialog, in case the user opened it via the "Tools > Build avalanche proof" menu and the utxos file contained unconfirmed coins. In such a case, the block height is shown in red in the utxos table, the coin is ignored when building the proof, a warning is shown to inform the user that some coins were not included, and an error is displayed in case all UTXOs have been excluded.

Test plan:
Create a few unconfirmed coins, some of them chained to also have unconfirmed parents. Check that if any of them is selected, the "Build avalanche proof" action is disabled in the drop down menu in the Coins tab.

Export a few coins, including these unconfirmed coins, in the Coins tab using the "Export coins detail" action. Try to use this file for building proofs through "Tools > Build avalanche proof" menu, and check that unconfirmed coins have their fake height shown in red. Click "Generate proof" and check that a warning dialog is shown to inform you that X coins have been excluded.

Repeat the previous test but only export unconfirmed coins to the JSON file, and check that now all coins have a red height, and that clicking "generate" in the proof building dialog causes an error to be shown to inform you that there are no coins left after excluding unconfirmed coins.