Network: Fix for uncorrect usage of python when formatting an errmsg
There was an incorrect usage of hfu(reversed(bytes)) in an errmsg which
was causing an exception. This path would only be taken if server is
serving wrong merkle branch.

This has been addressed.  Additionally, the verbose usage of
bytes(reversed(hfu(xxx))) has been modified to be the more readable and
less error-prone hfu(xx)[::-1].