refactor password prompting and caching in avaproof widget
The idea is to make it possible to reuse the password management code in the delegation widget, in order to be able to derive a deterministic delegated key using the wallet, if the user needs one.

A base widget is extracted from AvaProofWidget. This widget has a `pwd` property that can return a value set in the constructor or fetch the value interactively if needed. Once the password is set, it is remembered and no interaction is necessary for future access to `pwd`.

Test Plan:

Build a proof using a wallet without a password, and check that it works without ever prompting for a password.
Build a proof with an encrypted wallet. Try to cancel the password when prompted. Check that it is prompted again once when clicking "Generate".