Delegation dialog: provide a tool to generate a delegated key
Add a button to automatically generate a deterministic key for the delegated key. We use the wallet's seed with a derivation path that is not used for addresses (same technique as the automatic prefilling of the proof master key).

When the user clicks the button, an information dialog pops up telling the user to save the generated private key. When the popup is closed, the public key is used in the corresponding widget.

The use case for using a new key as a delegated key is when the owner of the proof and the owner of the delegation is the same user. He may want to create a delegation for his own proof so that the proof's private key does not need to be exposed on another machine (the one running the full node).