InvoiceDialog: add actions in the tools menu
Now the `if __name__ == '__main__'` block is no longer useful, it was only for testing the dialog before it was reachable through the application.

Also add a few typehints and set a minimum window size when opening a new invoice (so that the window is not resized if a different currency and API exchange rates need to be used).