InvoiceDialog: refactor ExchangeRateAPI class
Don't inherit the base class with child classes using different signatures. Just make the class use URLs and keys with placeholder strings for the currency, and replace the placeholder when needed for UI display or request purposes.

This will make the class usable for custom APIs specified by the user.